Fix : 회원탈퇴 오류수정
family가 null인경우 탈퇴가 되지않는 문제해결

+++ backend/src/main/java/geurime/api/service/UserServiceImpl.java
@@ -183,10 +183,12 @@ public class UserServiceImpl implements UserService {
                 .orElseThrow(() -> new CustomException(CustomExceptionList.USER_NOT_FOUND_ERROR));
 
         Family family = user.getFamily();
-        family.removeMember(user);
-        List<User> users = family.getUsers();
-        if(users.isEmpty()){
-            familyRepository.delete(family);
+        if(family != null){
+            family.removeMember(user);
+            List<User> users = family.getUsers();
+            if(users.isEmpty()){
+                familyRepository.delete(family);
+            }
         }
         userRepository.delete(user);
